backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
redis==5.0.8
celery==5.4.0
elasticsearch==8.14.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
pydantic==2.9.2
aiofiles==23.2.1
requests==2.32.3
python-multipart==0.0.12
itsdangerous==2.2.0
pyyaml==6.0.2

